chore: convert custom markdownlint rules to ESM
