+++ .gitignore
@@ -1,4 +1,2 @@
 dist
 node_modules
-markdownlint-rules/emd002.cjs
-markdownlint-rules/emd003.cjs
