--- markdownlint-rules/index.cjs
@@ -1,5 +1,0 @@
-const EMD002 = require('./emd002.cjs');
-const EMD003 = require('./emd003.cjs');
-const EMD004 = require('./emd004.cjs');
-
-module.exports = [EMD002, EMD003, EMD004];
