+++ markdownlint-rules/index.mjs
@@ -1,0 +1,5 @@
+import * as EMD002 from './emd002.mjs';
+import * as EMD003 from './emd003.mjs';
+import * as EMD004 from './emd004.mjs';
+
+export default [EMD002, EMD003, EMD004];
